% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_df_byid.R
\name{create_byperson_data}
\alias{create_byperson_data}
\title{Create a dataset with one gain per person (at the moment this selects the earliest gain), referred to as \code{byperson} dataset}
\usage{
create_byperson_data(data, bysg_data)
}
\arguments{
\item{data}{A dataset in wide format with an id variable and the sudden gains variables.}

\item{bysg_data}{A \code{bysg} dataset.}
}
\value{
A full dataset with all participants, the earliest gain per person
}
\description{
Create a dataset with one gain per person (at the moment this selects the earliest gain), referred to as \code{byperson} dataset
}
