% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_df_bysg.R
\name{create_df_bysg}
\alias{create_df_bysg}
\title{Create a dataset with one gain per row, referred to as \code{bysg} dataset}
\usage{
create_df_bysg(data, cutoff, id_var_name, sg_var_name)
}
\arguments{
\item{data}{A dataset in wide format with an id variable and the sudden gains variables.}

\item{cutoff}{A number specifying the cut-off for criterion 1.}

\item{id_var_name}{A string of the id variable name.}

\item{sg_var_name}{A string of the sudden gains variable name.}
}
\value{
A dataset with one row per sudden gain.
}
\description{
Create a dataset with one gain per row, referred to as \code{bysg} dataset
}
