% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_sg.R
\name{describe_sg}
\alias{describe_sg}
\title{Show descriptives for the sudden gains datasets}
\usage{
describe_sg(data)
}
\arguments{
\item{data}{A \code{bysg} or \code{byperson} dataset.}
}
\description{
Show descriptives for the sudden gains datasets
}
