% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_scores.R
\name{extract_scores}
\alias{extract_scores}
\title{Extract variables around the sudden gain}
\usage{
extract_scores(data, var_name, include_s0_extract = FALSE)
}
\arguments{
\item{data}{A \code{bysg} or \code{byperson} dataset.}

\item{var_name}{A string of the variable name which numbers numbers need to get extracted around the sudden gain.}
}
\value{
A wide dataset with values for \code{var_name} around the sudden gain.
}
\description{
Extract variables around the sudden gain
}
