% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_var.R
\name{extract_var_sg}
\alias{extract_var_sg}
\title{Extract variables around the sudden gain}
\usage{
extract_var_sg(data, var_name)
}
\arguments{
\item{data}{A \code{bysg} or \code{byperson} dataset.}

\item{var_name}{A string of the variable name which numbers numbers need to get extracted around the sudden gain.}
}
\value{
A wide dataset with values for \code{var_name} around the sudden gain.
}
\description{
Extract variables around the sudden gain
}
