% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_sg.R
\name{identify_sg}
\alias{identify_sg}
\title{Identify sudden gains in a wide dataset.}
\usage{
identify_sg(data, cutoff, id_var_name, sg_var_list,
  identify_sg_1to2 = FALSE, crit123_details = FALSE,
  name_crit123_details = "data_crit123_details")
}
\arguments{
\item{data}{A dataset in wide format with an id variable and the sudden gains variables.}

\item{cutoff}{A number specifying the cut-off for criterion 1.}

\item{id_var_name}{A string of the id variable name.}

\item{sg_var_list}{A LIST of the variable names for sudden gains analysis, if first session gains should be analysed simply add session 0 at the beginning of the list.}

\item{identify_sg_1to2}{Logical value to indicate whether first session gains should be in dataset,
simply include s0 before before all the other variables in sg_var_list, it will simply change the way
variables are named but this is super important for calculating which session sudden gain is in and also for extracting stuff later}

\item{crit123_details}{Logical, if crit123_details = TRUE a dataframe will be written to global environment}
}
\value{
A wide dataset indicating at which between session interval a sudden gain occured for each person in \code{data}.
}
\description{
Function identify sudden gains in a wide dataset. automatically checks whether first session gains are calculated by analysing the variable names looking for session with 0 (intake) before a letter.
}
