% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sg.R
\name{plot_sg_group}
\alias{plot_sg_group}
\title{Plot average magnitude of sudden gain or another variable around the sudden gain}
\usage{
plot_sg_group(data, ylabel, pre, post, start, pre3, pre2, pre1, post1, post2,
  post3, end)
}
\arguments{
\item{data}{A matched \code{byperson} dataset with variable sg_crit123 (0,1) as group variable.}

\item{ylabel}{A string as label for y-Axis.}

\item{start}{A string of the variable name with scores at the start of treatment.}

\item{pre3}{A string of the pre3 variable name.}

\item{pre2}{A string of the pre2 variable name.}

\item{pre1}{A string of the pre1 variable name.}

\item{post1}{A string of the post1 variable name.}

\item{post2}{A string of the post2 variable name.}

\item{post3}{A string of the post3 variable name.}

\item{end}{A string of the variable name with scores at the end of treatment.}
}
\value{
A plot of the average gain around or changes in scores of another questionnaire around the sudden gain.
}
\description{
Plot average magnitude of sudden gain or another variable around the sudden gain
}
