% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_sg_vars.R
\name{rename_sg_vars}
\alias{rename_sg_vars}
\title{Rename variables for extracting variables in sudden gians package}
\usage{
rename_sg_vars(data, rename_var_list, new_var_str = "temp_var_",
  start_numbering = 1)
}
\arguments{
\item{data}{Dataframe in wide format}

\item{rename_var_list}{List of variables to be ranamed}

\item{new_var_str}{TODO}

\item{start_numbering}{TODO}
}
\value{
Dataframe in wide format with renamed variables
}
\description{
Rename variables for extracting variables in sudden gians package
}
