% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sg_sample.R
\name{select_sg_sample}
\alias{select_sg_sample}
\title{Select sample with enough available data to identify sudden gains}
\usage{
select_sg_sample(data)
}
\arguments{
\item{data}{A dataset in wide format with an id variable and the sudden gains variables.}
}
\value{
A wide dataset a new variable \code{sg_select} indicating whether there is enough data available to identify sudden gains.
}
\description{
Select sample with enough available data to identify sudden gains
}
