% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdata.R
\docType{data}
\name{sgdata}
\alias{sgdata}
\title{Example dataset with BDI scores to illustrate how the package works}
\format{An longitudinal dataset in the wide format.}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(sgdata)
}
\description{
Generated dataset with made up BDI scores to illustrate how the package works
}
\examples{
# Load data into global environment
data(sgdata)
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{dataset}
