% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdata_bad.R
\docType{data}
\name{sgdata_bad}
\alias{sgdata_bad}
\title{Same as sgdata but with inconsistent, bad vaiable names}
\format{An longitudinal dataset in the wide format.}
\usage{
data(sgdata_bad)
}
\description{
To illistrate package also works with variable names dont have good names
}
\examples{
# Load data into global environment
data(sgdata_bad)
}
\keyword{dataset}
