% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinygains.R
\name{shinygains}
\alias{shinygains}
\title{This shiny application illustrates the main functions of this package}
\usage{
shinygains()
}
\value{

}
\description{
This shiny application illustrates the main functions of this package
}
\examples{
shinygains()
}
