% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_intervals.R
\name{count_intervals}
\alias{count_intervals}
\title{Number of between session intervals analysed for sudden gains}
\usage{
count_intervals(data, id_var_name, sg_var_list, identify_sg_1to2 = FALSE)
}
\arguments{
\item{data}{A data set in wide format including an ID variable and variables for each measurement point.}

\item{id_var_name}{String, specifying the name of the ID variable. Each row should have a unique value.}

\item{sg_var_list}{Vector, specifying the variable names of each measurement point sequentially.}

\item{identify_sg_1to2}{Logical, indicating whether to identify sudden losses from measurement point 1 to 2.
If set to TRUE, this implies that the first variable specified in \code{sg_var_list} represents a baseline measurement point, e.g. pre-intervention assessment.}
}
\value{
List with values for:
  \describe{
    \item{total_between_sess_intervals}{: The total number of between-session intervals present in the data set}
    \item{total_between_sess_intervals_sg}{: The total number of gain intervals (i.e. sudden gains) present in the data set}
    \item{sg_sess_sum_analysed}{: The total number of between-session intervals that could be analysed for sudden gains}
    \item{sg_sess_sum_not_analysed}{: The total number of between-session intervals that could not be analysed for sudden gains (due to missing data)}
    }
}
\description{
Calculates the number of between-session intervals present in the data set, the number that could feasibly be analysed for sudden gains (given the pattern of any missing data), and the number of gain intervals found.
}
