% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_intervals.R
\name{count_intervals}
\alias{count_intervals}
\title{Number of between session intervals analysed for sudden gains}
\usage{
count_intervals(data, cutoff, id_var_name, sg_var_list, analysed = TRUE,
  identify_sg_1to2 = FALSE)
}
\arguments{
\item{data}{A dataset in wide format with an id variable and the sudden gains variables.}

\item{cutoff}{A number specifying the cut-off for criterion 1.}

\item{id_var_name}{A string of the id variable name.}

\item{sg_var_list}{TODO.}

\item{analysed}{Logical, indicating whether sum of analysed or not analysed intervals should be reported. Analysed referrs to whether all three sudden gains criteria could be applied to a certain interval.}

\item{identify_sg_1to2}{Logical, indicating whether first session gains should be included, this is only possible if baseline scores are included in \code{sg_var_list}.}
}
\value{
Count of how many intervals were analysed for all three sudden gains criteria.
}
\description{
Number of between session intervals analysed for sudden gains
}
