% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_intervals.R
\name{count_intervals}
\alias{count_intervals}
\title{Number of between session intervals analysed for sudden gains}
\usage{
count_intervals(data, cutoff, id_var_name, sg_var_name, analysed = analyzed,
  analyzed = analysed)
}
\arguments{
\item{data}{A dataset in wide format with an id variable and the sudden gains variables.}

\item{cutoff}{A number specifying the cut-off for criterion 1.}

\item{id_var_name}{A string of the id variable name.}

\item{sg_var_name}{A string of the sudden gains variable name.}

\item{analysed}{A logical indicating wheter sum of analysed or not analysed session should be reported.}
}
\value{
A wide dataset indicating at which between session interval a sudden gain occured for each person in \code{data}.
}
\description{
Number of between session intervals analysed for sudden gains
}
