% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_intervals.R
\name{count_intervals}
\alias{count_intervals}
\title{Number of between session intervals analysed for sudden gains}
\usage{
count_intervals(data, sg_crit1_cutoff, id_var_name, sg_var_list,
  identify_sg_1to2 = FALSE)
}
\arguments{
\item{data}{A dataset in wide format with an id variable and the sudden gains variables.}

\item{sg_crit1_cutoff}{A number specifying the cut-off for criterion 1.}

\item{id_var_name}{A string of the id variable name.}

\item{sg_var_list}{TODO.}

\item{identify_sg_1to2}{Logical, indicating whether first session gains should be included, this is only possible if baseline scores are included in \code{sg_var_list}.}
}
\value{
List with values for:
total_between_sess_intervals,
total_between_sess_intervals_sg,
sg_sess_sum_analysed,
sg_sess_sum_not_analysed
}
\description{
I want  this to be different fomr title. Number of between session intervals analysed for sudden gains
}
