% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_crit1_cutoff.R
\name{define_crit1_cutoff}
\alias{define_crit1_cutoff}
\title{Define the Cut-off for SG Criterion 1 using the RCI Formula}
\usage{
define_crit1_cutoff(data_sessions, data_item)
}
\arguments{
\item{data_sessions}{A dataset in wide format with values for pre _s0 and post treatment _end}

\item{data_item}{A dataset in item-by-item scores for the SG measure}
}
\value{
A list with all calculated variables using formulate for RCI including cut-off
}
\description{
Define the Cut-off for SG Criterion 1 using the RCI Formula
}
