% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_sg.R
\name{describe_sg}
\alias{describe_sg}
\title{Show descriptives for the sudden gains datasets}
\usage{
describe_sg(data, sg_data_structure = c("bysg", "byperson"))
}
\arguments{
\item{data}{A \code{bysg} or \code{byperson} dataset.}

\item{sg_data_structure}{String indicating whether the provided data is a \code{bysg} or \code{byperson} dataset.}
}
\value{
todo
}
\description{
Some numbers (percentages) will be different depending wich dataset is selected, because if multiple gains in bysg only one will be selected for further analyses.
}
