% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_values.R
\name{extract_values}
\alias{extract_values}
\title{Extract variables around the sudden gain}
\usage{
extract_values(data, id_var_name, extract_var_list, extract_var_name,
  start_numbering = 1, add_to_data = TRUE)
}
\arguments{
\item{data}{A \code{bysg} or \code{byperson} dataset in wide format with the variable sg_session_n and all variables to extract.}

\item{id_var_name}{ID string, this is either normal id (for byperson datasets) or sg_id (for bysg datasets)}

\item{extract_var_list}{A string of the variable name which numbers numbers need to get extracted around the sudden gain.}

\item{extract_var_name}{TODO}

\item{start_numbering}{TODO}

\item{add_to_data}{TODO}
}
\value{
A wide dataset with values for \code{extract_var_name} around the sudden gain.
}
\description{
Extract variables around the sudden gain
}
