% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_values.R
\name{extract_values}
\alias{extract_values}
\title{Extract variables around the sudden gain}
\usage{
extract_values(data, id_var_name, extract_var_list, extract_measure_name,
  start_numbering = 1, add_to_data = TRUE)
}
\arguments{
\item{data}{A \code{bysg} or \code{byperson} dataset in wide format with the variable sg_session_n and all variables to extract.}

\item{id_var_name}{String, specifying the name of the ID variable. Each row should have a unique value.}

\item{extract_var_list}{Vector, specifying the variable names of session to session scores to extract from.}

\item{extract_measure_name}{String, specifying the name of the measure being used to extract values from.}

\item{start_numbering}{Numeric, set to by default 1.
Change to 0 if a pre-treatment (e.g. baseline assessment) measurement point is included in \code{extract_var_list}.}

\item{add_to_data}{Logical, if set to \code{TRUE}, the extracted values are added as new variables to the input dataset.
If set to false, only the ID variable and all extracted values will be returned.}
}
\value{
A wide dataset with values for \code{extract_measure_name} around the sudden gain.
}
\description{
Extract variables around the sudden gain
}
