% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_sg.R
\name{identify_sg}
\alias{identify_sg}
\title{Identify all sudden gains in a wide dataset.}
\usage{
identify_sg(data, cutoff, id_var_name, sg_var_name)
}
\arguments{
\item{data}{A dataset in wide format with an id variable and the sudden gains variables.}

\item{cutoff}{A number specifying the cut-off for criterion 1.}

\item{id_var_name}{A string of the id variable name.}

\item{sg_var_name}{A string of the sudden gains variable name.}
}
\value{
A wide dataset indicating at which between session interval a sudden gain occured for each person in \code{data}.
}
\description{
Identify all sudden gains in a wide dataset.
}
