% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_sl.R
\name{identify_sl}
\alias{identify_sl}
\title{Identify sudden losses.}
\usage{
identify_sl(data, id_var_name, sg_var_list, sg_crit1_cutoff,
  sg_crit2_pct = 0.25, identify_sg_1to2 = FALSE,
  crit123_details = FALSE,
  name_crit123_details = "data_crit123_details")
}
\arguments{
\item{data}{A data set in wide format including an ID variable and variables for each measurement point.}

\item{id_var_name}{String, specifying the name of the ID variable. Each row should have a unique value.}

\item{sg_var_list}{List, specifying the variable names of each measurement point sequentially.}

\item{sg_crit1_cutoff}{Numeric, specifying the negative cut-off value to be used for the first sudden losses criterion (see Examples below).}

\item{sg_crit2_pct}{Numeric, specifying the percentage change to be used for the second sudden losses criterion.}

\item{identify_sg_1to2}{Logical, indicating whether to identify sudden losses from measurement point 1 to 2.
If set to TRUE, this implies that the first variable specified in \code{sg_var_list} represents a baseline measurement point, e.g. pre-intervention assessment.}

\item{crit123_details}{Logical, if set to \code{TRUE} a dataframe will be wriiten to the global environment with information about which of the three criteria are met for each session to session interval for all cases.}

\item{name_crit123_details}{String, specifying the name of the dataframe produced if \code{crit123_details} is set to \code{TRUE}.}
}
\value{
A wide data set indicating whether sudden losses are present for each session to session interval for all cases in \code{data}.
}
\description{
Function to identify sudden losses in longitudinal data structured in wide format.
}
\examples{
identify_sl(data = sgdata,
            # Negative cut-off value to identify sudden losses
            sg_crit1_cutoff = -7,
            id_var_name = "id",
            sg_var_list = c("bdi_s1", "bdi_s2", "bdi_s3",
                            "bdi_s4", "bdi_s5", "bdi_s6",
                            "bdi_s7", "bdi_s8", "bdi_s9",
                            "bdi_s10", "bdi_s11", "bdi_s12"),
           identify_sg_1to2 = FALSE)
}
