% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sg.R
\name{plot_sg}
\alias{plot_sg}
\title{Plot average magnitude of sudden gain or another variable around the sudden gain}
\usage{
plot_sg(data, tx_start_var_name, tx_end_var_name, sg_pre_post_var_list,
  ylab, xlab = "Session")
}
\arguments{
\item{data}{A \code{bysg} or \code{byperson} dataset.}

\item{tx_start_var_name}{TODO}

\item{tx_end_var_name}{TODO}

\item{sg_pre_post_var_list}{TODO, list of all 3 variable names befor and 3 after the gain}

\item{ylab}{TODO}

\item{xlab}{TODO}
}
\value{
A plot of the average gain around or changes in scores of another questionnaire around the sudden gain.
}
\description{
Plot average magnitude of sudden gain or another variable around the sudden gain
}
