% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_sg_vars.R
\name{rename_sg_vars}
\alias{rename_sg_vars}
\title{Rename variables for extracting variables in sudden gians package}
\usage{
rename_sg_vars(data, sg_var_list, new_var_str = "temp_var_")
}
\arguments{
\item{data}{Dataframe in wide format}

\item{sg_var_list}{List of variables measuring first construct}
}
\value{
Dataframe in wide format with renamed variables
}
\description{
Rename variables for extracting variables in sudden gians package
}
