% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdata.R
\docType{data}
\name{sgdata}
\alias{sgdata}
\title{Example dataset with BDI scores and a fictional secondary process measure to illustrate how the package works}
\format{An longitudinal dataset in the wide format.}
\usage{
data(sgdata)
}
\description{
Some more detials here
}
\examples{
# Load data into global environment
data(sgdata)
}
\keyword{dataset}
