% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdata_bad.R
\docType{data}
\name{sgdata_bad}
\alias{sgdata_bad}
\title{Longitudinal dataset with repeated measures of depression and rumination (bad variable names)}
\format{A longitudinal dataset in wide format:
\itemize{
  \item{id}{: ID variable, unique identifier for each person}
  \item{bdi_intake}{: BDI value, baseline assessment}
  \item{bdi1}{: BDI value, session 1}
  \item{wk2bdi}{: BDI value, session 2}
  \item{bdi.S3}{: BDI value, session 3}
  \item{bdis4}{: BDI value, session 4}
  \item{bdi_session5}{: BDI value, session 5}
  \item{bdi_weekSix}{: BDI value, session 6}
  \item{beck.dep.inv_sess7}{: BDI value, session 7}
  \item{weeeek8bdi}{: BDI value, session 8}
  \item{bdi_nine}{: BDI value, session 9}
  \item{bdii10}{: BDI value, session 10}
  \item{bDi11}{: BDI value, session 11}
  \item{bdi_s12_end}{: BDI value, session 12}
  \item{bdi_fu1}{: BDI value, follow-up measure 1}
  \item{bdi_followup2}{: BDI value, follow-up measure 2}
  \item{rq_intake}{: RQ value, baseline assessment}
  \item{rqi1}{: RQ value, session 1}
  \item{wk2rq}{: RQ value, session 2}
  \item{rq.S3}{: RQ value, session 3}
  \item{rq_s4}{: RQ value, session 4}
  \item{rq_session5}{: RQ value, session 5}
  \item{rq_weekSix}{: RQ value, session 6}
  \item{rq_sess7}{: RQ value, session 7}
  \item{weeeek8rq}{: RQ value, session 8}
  \item{rqnine}{: RQ value, session 9}
  \item{rq10}{: RQ value, session 10}
  \item{rqi11}{: RQ value, session 11}
  \item{rq_s12_end}{: RQ value, session 12}
  \item{prq_fu1}{: RQ value, follow-up measure 1}
  \item{rqq_followup2}{: RQ value, follow-up measure 2}
}}
\usage{
data(sgdata_bad)
}
\description{
Same as \code{sgdata} but with bad vaiable names to illustrate that variable names can be inconsistent.
}
\examples{
# Load data into global environment
data(sgdata_bad)
}
\keyword{dataset}
